+++ src/module/vue/components/progress/progress-box.vue
@@ -106,15 +106,12 @@ export default {
       return actor?.items.get(this.item._id)
     },
     subtitle() {
-      let subtype=this.$capitalize(this.item.data.subtype)
+      let subtype = this.$capitalize(this.item.data.subtype)
       if (subtype === 'Bond') subtype = 'Connection' // translate name
       return this.$t(`IRONSWORN.${subtype}`)
     },
     completedIcon() {
-      const suffix = this.item.data.completed
-        ? 'fa-check-circle'
-        : 'fa-dot-circle'
-      return `fab ${suffix}`
+      return this.item.data.completed ? 'check-circle' : 'circle-notch'
     },
     completedTooltip() {
       const suffix = this.item.data.completed ? 'Completed' : 'NotCompleted'